requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
httpx
google-cloud-aiplatform
